+++ R/phi_cramers_v.R
@@ -1,24 +1,30 @@
 #' Effect size for contingency tables
 #'
-#' Compute Cramer's V and phi (\eqn{\phi}) from contingency tables.
+#' Compute Cramer's V and phi (\eqn{\phi}) for contingency tables or
+#' goodness-of-fit.
 #'
-#' @param x a numeric vector or matrix. x and y can also both be factors.
-#' @param y a numeric vector; ignored if x is a matrix. If x is a factor, y should be a factor of the same length.
+#' @inheritParams stats::chisq.test
 #' @param ci Confidence Interval (CI) level
 #' @param adjust Should the effect size be bias-corrected? Defaults to `FALSE`.
 #' @param CI Deprecated in favor of `ci`.
-#' @param ... Ignored.
+#' @param ... Arguments passed to [chisq.test()], such as `p`.
 #'
 #' @return A data frame with the effect size(s) between 0-1, and confidence interval(s).
 #'
 #' @seealso [chisq_to_phi()] for details regarding estimation and CIs.
 #'
 #' @examples
-#' contingency_table <- as.table(rbind(c(762, 327, 468), c(484, 239, 477), c(484, 239, 477)))
+#' M <- rbind(c(150, 130, 35, 55),
+#'            c(100, 50,  10, 40),
+#'            c(165, 65,  2,  25))
 #'
-#' phi(contingency_table)
+#' dimnames(M) <- list(Study = c("Psych", "Econ", "Law"),
+#'                     Music = c("Pop", "Rock", "Jazz", "Classic"))
+#' M
 #'
-#' cramers_v(contingency_table)
+#' phi(M)
+#'
+#' cramers_v(M)
 #'
 #' @importFrom stats chisq.test
 #' @export
@@ -28,7 +34,7 @@ phi <- function(x, y = NULL, ci = 0.95, adjust = FALSE, CI, ...){
     warning("'CI' argument is deprecated. Use 'ci' instead.")
   }
 
-  res <- suppressWarnings(stats::chisq.test(x, y))
+  res <- suppressWarnings(stats::chisq.test(x, y, ...))
   Obs <- res$observed
   Exp <- res$expected
 
@@ -49,7 +55,7 @@ cramers_v <- function(x, y = NULL, ci = 0.95, adjust = FALSE, CI,...){
     warning("'CI' argument is deprecated. Use 'ci' instead.")
   }
 
-  res <- suppressWarnings(stats::chisq.test(x, y))
+  res <- suppressWarnings(stats::chisq.test(x, y, ...))
   Obs <- res$observed
   Exp <- res$expected